tf-lite-pb-onnx-detection/requirements.txt
numpy==1.16.3
opencv-python==4.1.0.25
tensorflow-gpu==2.0.0